Switched build system to autotools.
This commit switches libtweetnacl from a static Makefile to GNU
autotools as the build system. There will probably be a few
commits coming down the pipeline that extend/tweak the build-system
behavior.

A notable difference is that the data source for randombytes() is now
selected by the user during ./configure instead of depending on a
manually-created header file.