Add automake support for producing test wrappers.
Add support for producing tweetnacl and libsodium test-wrapper
libraries. Note that a little bit more work will be needed to get
randombytes() into the tweetnacl version.

These new libraries pave the way for the `make check` tests that will
be coming in future commits as we move towards a full test-suite that
compares tweetnacl against libsodium.